requirements.txt
alembic==1.8.0
amqp==5.1.1
anyio==3.6.1
apache-airflow==2.3.3
apache-airflow-providers-amazon==5.0.0
apache-airflow-providers-celery==3.0.0
apache-airflow-providers-common-sql==1.0.0
apache-airflow-providers-ftp==3.0.0
apache-airflow-providers-http==3.0.0
apache-airflow-providers-imap==3.0.0
apache-airflow-providers-mongo==3.0.0
apache-airflow-providers-snowflake==3.1.0
apache-airflow-providers-sqlite==3.0.0
apispec==3.3.2
argcomplete==2.0.0
asn1crypto==1.5.1
attrs==20.3.0
Babel==2.10.3
beautifulsoup4==4.11.1
billiard==3.6.4.0
blinker==1.4
boto3==1.24.51
botocore==1.27.51
cachelib==0.9.0
cattrs==1.10.0
celery==5.2.7
certifi==2020.12.5
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clickclick==20.10.2
colorama==0.4.5
colorlog==4.8.0
commonmark==0.9.1
connexion==2.14.0
cramjam==2.5.0
cron-descriptor==1.2.30
croniter==1.3.5
cryptography==36.0.2
decorator==5.1.1
Deprecated==1.2.13
dill==0.3.1.1
dnspython==2.2.1
docutils==0.19
email-validator==1.2.1
fastparquet==0.8.1
Flask==2.1.2
Flask-AppBuilder==4.1.2
Flask-Babel==2.0.0
Flask-Caching==2.0.0
Flask-JWT-Extended==4.4.2
Flask-Login==0.6.1
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
flower==1.1.0
fsspec==2022.7.1
graphviz==0.20
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
humanize==4.2.3
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpath-ng==1.5.3
jsonschema==4.6.1
kombu==5.2.4
lazy-object-proxy==1.7.1
linkify-it-py==2.0.0
lockfile==0.12.2
lxml==4.9.1
Mako==1.2.1
Markdown==3.3.7
markdown-it-py==2.1.0
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.3.0
mdurl==0.1.1
mypy-boto3-appflow==1.24.36.post1
mypy-boto3-rds==1.24.51
mypy-boto3-redshift-data==1.24.36.post1
numpy==1.23.2
oscrypto==1.3.0
packaging==21.3
pandas==1.4.3
pathspec==0.9.0
pendulum==2.1.2
pkg_resources==0.0.0
pluggy==1.0.0
ply==3.11
prison==0.2.1
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
pyarrow==8.0.0
pycparser==2.21
pycryptodomex==3.15.0
Pygments==2.12.0
PyJWT==2.4.0
pymongo==3.6.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-daemon==2.3.0
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==6.1.2
pytz==2022.1
pytzdata==2020.1
PyYAML==6.0
redshift-connector==2.0.908
requests==2.28.0
rfc3986==1.5.0
rich==12.4.4
s3transfer==0.6.0
scramp==1.4.1
setproctitle==1.2.3
six==1.16.0
sniffio==1.2.0
snowflake-connector-python==2.7.11
snowflake-sqlalchemy==1.3.4
soupsieve==2.3.2.post1
SQLAlchemy==1.4.27
SQLAlchemy-JSONField==1.0.0
sqlalchemy-redshift==0.8.11
SQLAlchemy-Utils==0.38.2
swagger-ui-bundle==0.0.9
tabulate==0.8.10
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
tornado==6.2
typing_extensions==4.3.0
uc-micro-py==1.0.1
unicodecsv==0.14.1
urllib3==1.26.9
vine==5.0.0
watchtower==2.0.1
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
WTForms==2.3.3
zipp==3.8.0
